net80211: fix possible overflow in IEEE80211_TU_TO_TICKS()
For hz=1000 any number, greater than 4194 causes integer overflow;
this change casts the number to uint64_t before operating with it.

Approved by:	adrian (mentor)
Differential Revision:	https://reviews.freebsd.org/D5268


Former-commit-id: 8276e07f95843e55289dd8b64450b991432b9b16

+++ sys/net80211/ieee80211_var.h
@@ -377,7 +377,7 @@ name|IEEE80211_TU_TO_TICKS
 parameter_list|(
 name|x
 parameter_list|)
-value|(((x) * 1024 * hz) / (1000 * 1000))
+value|(((uint64_t)(x) * 1024 * hz) / (1000 * 1000))
 end_define
 
 begin_comment
